+++ src/main/java/kr/co/pawong/pwbe/chat/adapter/in/api/ChatMessageCommandController.java
@@ -1,7 +1,7 @@
 package kr.co.pawong.pwbe.chat.adapter.in.api;
 
 import kr.co.pawong.pwbe.chat.application.port.in.CommandChatMessageDataUseCase;
-import kr.co.pawong.pwbe.user.adapter.out.security.CustomUserDetails;
+import kr.co.pawong.pwbe.global.security.dto.CustomUserDetails;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
